Rough draft of xHCI trace events.
This is the start of a rough outline of where I would like to see the
trace events project go.  There are probably still more events to
define, and I may reorganize them a bit after seeing how my debugging
flow works with turning on and off specific trace events.

Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>